+++ src/hooks/xp-rule-checker.js
